chore(j-s): Test Improvements (#19654)
* Fixes various memory leak issues in unit tests

* Refactors code

* Fixes barrel file exports

* Update apps/judicial-system/web/src/components/InfoCard/DefendantInfo/DefendantInfo.logic.ts


* Fixes spelling

* Fixes formatting

---------
