+++ apps/judicial-system/web/jest.config.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 export default {
   preset: './jest.preset.js',
   rootDir: '../../..',
