+++ apps/judicial-system/web/src/components/BlueBoxWithIcon/BlueBoxWithDate.tsx
@@ -22,7 +22,7 @@ import {
 import { formatDateForServer, useDefendants } from '../../utils/hooks'
 import DateTime from '../DateTime/DateTime'
 import { FormContext } from '../FormProvider/FormProvider'
-import { getAppealExpirationInfo } from '../InfoCard/DefendantInfo/DefendantInfo'
+import { getAppealExpirationInfo } from '../InfoCard/DefendantInfo/DefendantInfo.logic'
 import SectionHeading from '../SectionHeading/SectionHeading'
 import { strings } from './BlueBoxWithDate.strings'
 import * as styles from './BlueBoxWithIcon.css'
