+++ apps/judicial-system/web/src/components/Table/Table.tsx
@@ -27,7 +27,7 @@ import {
 } from '../../types'
 import { useCase, useCaseList, useViewport } from '../../utils/hooks'
 import { compareLocaleIS } from '../../utils/sortHelper'
-import { mapCaseStateToTagVariant } from '../Tags/TagCaseState/TagCaseState'
+import { mapCaseStateToTagVariant } from '../Tags/TagCaseState/TagCaseState.logic'
 import DurationDate, { getDurationDate } from './DurationDate/DurationDate'
 import SortButton from './SortButton/SortButton'
 import { table as strings } from './Table.strings'
