+++ apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/Components/ModifyDatesModal/ModifyDatesModal.logic.spec.ts
@@ -1,16 +1,15 @@
-import { createIntl } from 'react-intl'
+import { createFormatMessage } from '@island.is/judicial-system-web/src/utils/testHelpers.logic'
 
-import { createCaseModifiedExplanation } from './ModifyDatesModal'
+import { createCaseModifiedExplanation } from './ModifyDatesModal.logic'
 
 describe('createCaseModifiedExplanation', () => {
-  const formatMessage = createIntl({
-    locale: 'is',
-    onError: jest.fn(),
-  }).formatMessage
-  beforeAll(() => jest.useFakeTimers())
+  const formatMessage = createFormatMessage()
+
+  beforeEach(() => jest.useFakeTimers())
+  afterEach(() => jest.useRealTimers())
 
   it('should append nextExplainantion', () => {
-    const previousExplaination =
+    const previousExplanation =
       'Föstud. 10. júní 2022 kl. 10:29 - Test Testsen dómari, Héraðsdómur Reykjavíkur<br/>Ástæða: Testing'
     const nextExplainantion = 'Testing2'
     const userName = 'Test Testsen'
@@ -21,7 +20,7 @@ describe('createCaseModifiedExplanation', () => {
 
     const res = createCaseModifiedExplanation(
       formatMessage,
-      previousExplaination,
+      previousExplanation,
       nextExplainantion,
       userName,
       userTitle,
@@ -33,8 +32,8 @@ describe('createCaseModifiedExplanation', () => {
     )
   })
 
-  it('should format correctly without previousExplaination', () => {
-    const previousExplaination = undefined
+  it('should format correctly without previousExplanation', () => {
+    const previousExplanation = undefined
     const nextExplainantion = 'Testing'
     const userName = 'Test Testsen'
     const userTitle = 'Dómari'
@@ -44,7 +43,7 @@ describe('createCaseModifiedExplanation', () => {
 
     const res = createCaseModifiedExplanation(
       formatMessage,
-      previousExplaination,
+      previousExplanation,
       nextExplainantion,
       userName,
       userTitle,
