+++ apps/judicial-system/web/src/utils/testHelpers.logic.ts
@@ -1,0 +1,7 @@
+import { createIntl } from 'react-intl'
+
+export const createFormatMessage = () =>
+  createIntl({
+    locale: 'is',
+    onError: jest.fn,
+  }).formatMessage
