+++ apps/judicial-system/web/src/utils/testHelpers.tsx
@@ -1,15 +1,10 @@
 import { FC, PropsWithChildren, ReactNode } from 'react'
-import { createIntl, IntlProvider } from 'react-intl'
+import { IntlProvider } from 'react-intl'
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client'
 
 import { FormContext, UserContext } from '../components'
 import { Case, UserRole } from '../graphql/schema'
 import { mockUser } from './mocks'
-
-export const formatMessage = createIntl({
-  locale: 'is',
-  onError: jest.fn,
-}).formatMessage
 
 export const IntlProviderWrapper: FC<PropsWithChildren> = ({ children }) => {
   return (
