+++ apps/judicial-system/web/src/utils/utils.spec.ts
@@ -1,7 +1,6 @@
 import faker from 'faker'
 
 import {
-  Case,
   Defendant,
   Gender,
   Notification,
@@ -9,7 +8,6 @@ import {
 } from '@island.is/judicial-system-web/src/graphql/schema'
 
 import * as formatters from './formatters'
-import { validateAndSendToServer } from './formHelper'
 import {
   getDefaultDefendantGender,
   getShortGender,
@@ -244,53 +242,6 @@ describe('Utils', () => {
 
       // Assert
       expect(res).toEqual(true)
-    })
-  })
-
-  describe('validateAndSendToServer', () => {
-    test('should call the updateCase function with the correct parameters', () => {
-      // Arrange
-      const spy = jest.fn()
-      const fieldToUpdate = 'courtCaseNumber'
-      const value = '1234/1234'
-      const id = faker.datatype.uuid()
-      const theCase = { id } as Case
-      const update = {
-        courtCaseNumber: value,
-      }
-
-      // Act
-      validateAndSendToServer(
-        fieldToUpdate,
-        value,
-        ['appeal-case-number-format'],
-        theCase,
-        spy,
-      )
-
-      // Assert
-      expect(spy).toBeCalledWith(id, update)
-    })
-
-    test('should not call the updateCase function if the value is invalid', () => {
-      // Arrange
-      const spy = jest.fn()
-      const fieldToUpdate = 'courtCaseNumber'
-      const value = '12341234'
-      const id = faker.datatype.uuid()
-      const theCase = { id } as Case
-
-      // Act
-      validateAndSendToServer(
-        fieldToUpdate,
-        value,
-        ['appeal-case-number-format'],
-        theCase,
-        spy,
-      )
-
-      // Assert
-      expect(spy).not.toHaveBeenCalled()
     })
   })
