+++ apps/judicial-system/web/src/utils/validate.ts
@@ -22,7 +22,6 @@ import {
   Victim,
 } from '@island.is/judicial-system-web/src/graphql/schema'
 
-import { Updates } from '../routes/Court/Indictments/Subpoena/Subpoena'
 import { isBusiness } from './utils'
 
 export type Validation =
@@ -45,66 +44,59 @@ const getRegexByValidation = (validation: Validation) => {
   switch (validation) {
     case 'empty':
       return {
-        regex: new RegExp('.'),
+        regex: /./,
         errorMessage: 'Reitur má ekki vera tómur',
       }
     case 'time-format':
       return {
-        regex: new RegExp(/^((0[0-9]|1[0-9]|2[0-3])|[0-9]):[0-5][0-9]$/),
+        regex: /^((0[0-9]|1[0-9]|2[0-3])|[0-9]):[0-5][0-9]$/,
         errorMessage: 'Dæmi: 12:34 eða 1:23',
       }
     case 'police-casenumber-format':
       return {
-        regex: new RegExp(/^[0-9][0-9][0-9]-[0-9][0-9][0-9][0-9]-\d{1,99999}$/),
+        regex: /^[0-9][0-9][0-9]-[0-9][0-9][0-9][0-9]-\d{1,99999}$/,
         errorMessage: 'Dæmi: 012-3456-7890',
       }
     case 'national-id':
       return {
-        regex: new RegExp(/^\d{6}(-?\d{4})?$/),
+        regex: /^\d{6}(-?\d{4})?$/,
         errorMessage: 'Dæmi: 000000-0000',
       }
-    case 'date-of-birth': {
+    case 'date-of-birth':
       return {
-        regex: new RegExp(
-          /^(0[1-9]|[12][0-9]|3[01])[.](0[1-9]|1[012])[.](19|20)\d\d$/,
-        ),
+        regex: /^(0[1-9]|[12][0-9]|3[01])[.](0[1-9]|1[012])[.](19|20)\d\d$/,
         errorMessage: 'Dæmi: 00.00.0000',
       }
-    }
     case 'email-format':
       return {
-        regex: new RegExp(/^[\w-.+]+@([\w-]+\.)+[\w-]{2,4}$/),
+        regex: /^[\w-.+]+@([\w-]+\.)+[\w-]{2,4}$/,
         errorMessage: 'Netfang ekki á réttu formi',
       }
     case 'phonenumber':
       return {
-        regex: new RegExp(/^(\d{3})(-?(\d{4}))$/),
+        regex: /^(\d{3})(-?(\d{4}))$/,
         errorMessage: 'Dæmi: 555-5555',
       }
-    case 'date-format': {
+    case 'date-format':
       return {
-        regex: new RegExp(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(.\d{3})?Z$/),
+        regex: /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(.\d{3})?Z$/,
         errorMessage: '',
       }
-    }
-    case 'R-case-number': {
+    case 'R-case-number':
       return {
-        regex: new RegExp(/^R-[0-9]{1,5}\/[0-9]{4}$/),
+        regex: /^R-[0-9]{1,5}\/[0-9]{4}$/,
         errorMessage: `Dæmi: R-1234/${new Date().getFullYear()}`,
       }
-    }
-    case 'S-case-number': {
+    case 'S-case-number':
       return {
-        regex: new RegExp(/^S-[0-9]{1,5}\/[0-9]{4}$/),
+        regex: /^S-[0-9]{1,5}\/[0-9]{4}$/,
         errorMessage: `Dæmi: S-1234/${new Date().getFullYear()}`,
       }
-    }
-    case 'appeal-case-number-format': {
+    case 'appeal-case-number-format':
       return {
-        regex: new RegExp(/^[0-9]{1,4}\/[0-9]{4}$/),
+        regex: /^[0-9]{1,4}\/[0-9]{4}$/,
         errorMessage: `Dæmi: 1234/${new Date().getFullYear()}`,
       }
-    }
   }
 }
 
@@ -127,10 +119,12 @@ export const validate = (items: ValidateItem[]): IsValid => {
             return undefined
           }
 
-          const v = getRegexByValidation(validation)
+          const config = getRegexByValidation(validation)
 
-          const isValid = v.regex.test(value)
-          return isValid ? undefined : { isValid, errorMessage: v.errorMessage }
+          const isValid = config.regex.test(value)
+          return isValid
+            ? undefined
+            : { isValid, errorMessage: config.errorMessage }
         })
         .filter(Boolean) as IsValid[]
 
@@ -516,10 +510,11 @@ export const isCourtRecordStepValidIC = (workingCase: Case): boolean => {
 
 export const isSubpoenaStepValid = (
   workingCase: Case,
-  updates?: Updates,
+  updatedDefendants?: Defendant[] | null,
+  updatedArraignmentDate?: DateLog | null,
 ): boolean => {
-  const { arraignmentDate } = updates?.theCase || workingCase
-  const { defendants } = updates || workingCase
+  const arraignmentDate = updatedArraignmentDate || workingCase.arraignmentDate
+  const defendants = updatedDefendants || workingCase.defendants
 
   const validateDefendants = (defendants?: Defendant[] | null) => {
     return defendants?.every((defendant) =>
